requirements.txt
flask
python-dotenv
requests
Pillow
openpyxl
gunicorn
gevent